Refactor Indonesian researcher
