+++ packages/yoastseo/spec/morphology/indonesian/stemSpec.js
@@ -1,4 +1,4 @@
-import stem from "../../../src/languageProcessing/languages/id/morphology/stem";
+import stem from "../../../src/languageProcessing/languages/id/helpers/internal/stem";
 import getMorphologyData from "../../specHelpers/getMorphologyData";
 
 const morphologyData = getMorphologyData( "id" ).id;
