+++ packages/yoastseo/src/languageProcessing/languages/id/config/internal/nonPassiveVerbsStartingDi.js
