+++ packages/yoastseo/src/languageProcessing/languages/id/helpers/internal/stemHelpers.js
