--- packages/yoastseo/src/languageProcessing/languages/id/researches/getKeywordDensity.js
@@ -1,10 +1,0 @@
-import getKeywordDensity from "../../../researches/base/getKeywordDensity";
-
-/**
- * Calculates the keyword density and returns the information whether a stemmer is available or not.
- *
- * @inheritDoc getKeywordDensity
- */
-export default function( paper, researcher ) {
-	return getKeywordDensity( paper, researcher, true );
-}
